Extract ANSI rendering logic from console
