+++ src/vs/workbench/browser/positronAnsiRenderer/outputLine.tsx
@@ -5,11 +5,15 @@
 import 'vs/css!./outputLine';
 import * as React from 'react';
 import { ANSIOutputLine } from 'vs/base/common/ansiOutput';
-import { OutputRun } from 'vs/workbench/contrib/positronConsole/browser/components/outputRun';
+import { OutputRun } from 'vs/workbench/browser/positronAnsiRenderer/outputRun';
+import { IOpenerService } from 'vs/platform/opener/common/opener';
+import { INotificationService } from 'vs/platform/notification/common/notification';
 
 // OutputLineProps interface.
 export interface OutputLineProps {
 	readonly outputLine: ANSIOutputLine;
+	readonly openerService: IOpenerService;
+	readonly notificationService: INotificationService;
 }
 
 /**
@@ -18,13 +22,19 @@ export interface OutputLineProps {
  * @returns The rendered component.
  */
 export const OutputLine = (props: OutputLineProps) => {
+
 	// Render.
 	return (
 		<div>
 			{!props.outputLine.outputRuns.length ?
 				<br /> :
 				props.outputLine.outputRuns.map(outputRun =>
-					<OutputRun key={outputRun.id} outputRun={outputRun} />
+					<OutputRun
+						key={outputRun.id}
+						outputRun={outputRun}
+						openerService={props.openerService}
+						notificationService={props.notificationService}
+					/>
 				)
 			}
 		</div>
