+++ src/vs/workbench/browser/positronAnsiRenderer/outputLines.css
