+++ src/vs/workbench/browser/positronAnsiRenderer/outputRun.css
