+++ src/vs/workbench/contrib/positronConsole/browser/components/activityInput.tsx
@@ -8,8 +8,9 @@ import { useState } from 'react'; // eslint-disable-line no-duplicate-imports
 import { DisposableStore } from 'vs/base/common/lifecycle';
 import { FontInfo } from 'vs/editor/common/config/fontInfo';
 import { positronClassNames } from 'vs/base/common/positronUtilities';
-import { OutputRun } from 'vs/workbench/contrib/positronConsole/browser/components/outputRun';
 import { ActivityItemInput, ActivityItemInputState } from 'vs/workbench/services/positronConsole/browser/classes/activityItemInput';
+import { usePositronConsoleContext } from 'vs/workbench/contrib/positronConsole/browser/positronConsoleContext';
+import { OutputRun } from 'vs/workbench/browser/positronAnsiRenderer/outputRun';
 
 // ActivityInputProps interface.
 export interface ActivityInputProps {
@@ -26,6 +27,9 @@ export const ActivityInput = (props: ActivityInputProps) => {
 	// Hooks.
 	const [state, setState] = useState(props.activityItemInput.state);
 
+	// Get services from the context.
+	const { openerService, notificationService } = usePositronConsoleContext();
+
 	// Main useEffect.
 	React.useEffect(() => {
 		// Create the disposable store for cleanup.
@@ -38,7 +42,7 @@ export const ActivityInput = (props: ActivityInputProps) => {
 
 		// Return the cleanup function that will dispose of the disposables.
 		return () => disposableStore.dispose();
-	}, []);
+	}, [props.activityItemInput]);
 
 	// Calculate the prompt length.
 	const promptLength = Math.max(
@@ -68,7 +72,12 @@ export const ActivityInput = (props: ActivityInputProps) => {
 						}
 					</span>
 					{outputLine.outputRuns.map(outputRun =>
-						<OutputRun key={outputRun.id} outputRun={outputRun} />
+						<OutputRun
+							key={outputRun.id}
+							outputRun={outputRun}
+							notificationService={notificationService}
+							openerService={openerService}
+						/>
 					)}
 				</div>
 			)}
